@@ -1,19 +1,16 @@
 class UsersController < ApplicationController
   before_filter :login_required
 
-  # render new.rhtml
   def new
     @user = User.new
   end
 
   def create
     @user = User.new(params[:user])
-    success = @user && @user.save
-    if success && @user.errors.empty?
-      redirect_to('/')
+    if @user.save
+      redirect_to(root_path)
       flash[:notice] = "User created."
     else
-      flash[:error]  = "Errors; please try again."
       render :action => 'new'
     end
   end
