@@ -1,9 +1,9 @@
 class CiMonitorMailer < ActionMailer::Base
+  default :from => SYSTEM_ADMIN_EMAIL
 
   def red_over_one_day_notification(projects, options = {})
-    from("Pivotal CiMonitor <pivotal-cimonitor@example.com>")
-    recipients(RED_NOTIFICATION_EMAILS)
-    subject("Projects RED for over one day!")
-    multipart("red_over_one_day_notification", :projects => projects)
+    @projects = projects
+    mail :to => RED_NOTIFICATION_EMAILS,
+         :subject => "Projects RED for over one day!"
   end
 end
