@@ -3,6 +3,6 @@ class CiMonitorNotifier
     projects = Project.find(:all, :conditions => {:enabled => true}).select do |project|
       project.red_since && project.red_since < Clock.now - 1.day
     end
-    CiMonitorMailer.deliver_red_over_one_day_notification(projects) unless projects.empty?
+    CiMonitorMailer.red_over_one_day_notification(projects).deliver unless projects.empty?
   end
 end
