@@ -1,3 +1,6 @@
+require 'openid'
+require 'openid/extensions/ax'
+
 class User < ActiveRecord::Base
   include Authentication
   include Authentication::ByPassword
@@ -24,10 +27,30 @@ class User < ActiveRecord::Base
   end
 
   def login=(value)
-    write_attribute :login, (value ? value.downcase : nil)
+    self[:login] = value && value.downcase
   end
 
   def email=(value)
-    write_attribute :email, (value ? value.downcase : nil)
+    self[:email] = value && value.downcase
   end
+
+  def self.find_or_create_from_google_openid(fetch_response)
+    email = fetch_response.get_single('http://axschema.org/contact/email')
+    first_name = fetch_response.get_single('http://axschema.org/namePerson/first')
+    last_name = fetch_response.get_single('http://axschema.org/namePerson/last')
+
+    email_parts = email.split('@')
+    login = email_parts.first
+
+    user = User.find_by_login(login) || User.new(:login => login)
+    full_name = "#{first_name} #{last_name}"
+    user.name = full_name.blank? ? "" : full_name
+    user.email = email
+
+    # todo - this is a bit of a hack for now...
+    user.password = user.password_confirmation = ActiveSupport::SecureRandom.hex(16)
+    user.save!
+    user
+  end
+
 end
