@@ -5,9 +5,9 @@ class ActionMailer::Base
     content_type 'multipart/alternative'
 
     part :content_type => "text/plain",
-         :body => render_message("#{template_name}.plain.erb", assigns)
+         :body => render("#{template_name}.plain.erb", assigns)
 
     part :content_type => "text/html",
-         :body => render_message("#{template_name}.html.erb", assigns)
+         :body => render("#{template_name}.html.erb", assigns)
   end
 end
