@@ -2,9 +2,8 @@
 require 'fileutils'
 
 Project.configure do |project|
-  FileUtils.cp File.join(project.path, 'work', 'config', 'database.yml.example'), File.join(project.path, 'work', 'config', 'database.yml')
-  FileUtils.cp File.join(project.path, 'work', 'config', 'initializers', 'site_keys.rb.example'), File.join(project.path, 'work', 'config', 'initializers', 'site_keys.rb')
+#  FileUtils.cp File.join(project.path, 'work', 'config', 'database.yml.example'), File.join(project.path, 'work', 'config', 'database.yml')
   project.email_notifier.emails = ["pivotal-cimonitor@example.com"]
   ENV['CRUISE_PROJECT_NAME'] = project.name
-  project.build_command = 'ruby ci/ci_build.rb'
+  project.build_command = './cruise_build.sh'
 end
