@@ -54,7 +54,7 @@ module AuthenticatedSystem
   end
 
   def not_logged_in_url
-    return new_login_path
+    return login_path
   end
 
   # Redirect as appropriate when an access request fails.
@@ -85,7 +85,7 @@ module AuthenticatedSystem
   #
   # We can return to this location by calling #redirect_back_or_default.
   def store_location
-    session[:return_to] = request.request_uri
+    session[:return_to] = request.fullpath
   end
 
   # Redirect to the URI stored by the most recent store_location call or
