@@ -8,7 +8,7 @@ spec_builder = Jasmine::SpecBuilder.new(jasmine_config)
 
 should_stop = false
 
-Spec::Runner.configure do |config|
+RSpec.configure do |config|
   config.after(:suite) do
     spec_builder.stop if should_stop
   end
