@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../spec_helper'
+require 'spec_helper'
 
 describe User do
   fixtures :users
@@ -20,17 +20,17 @@ describe User do
   it 'requires login' do
     lambda do
       u = create_user(:login => nil)
-      u.errors.on(:login).should_not be_nil
+      u.errors[:name].should_not be_nil
     end.should_not change(User, :count)
   end
 
   describe 'allows legitimate logins:' do
     ['123', '1234567890_234567890_234567890_234567890',
      'hello.-_there@funnychar.com'].each do |login_str|
-      it "'#{login_str}'" do
+      it "should work for '#{login_str}'" do
         lambda do
           u = create_user(:login => login_str)
-          u.errors.on(:login).should be_nil
+          u.errors[:name].should be_empty
         end.should change(User, :count).by(1)
       end
     end
@@ -42,7 +42,7 @@ describe User do
       it "'#{login_str}'" do
         lambda do
           u = create_user(:login => login_str)
-          u.errors.on(:login).should_not be_nil
+          u.errors[:name].should_not be_nil
         end.should_not change(User, :count)
       end
     end
@@ -51,21 +51,21 @@ describe User do
   it 'requires password' do
     lambda do
       u = create_user(:password => nil)
-      u.errors.on(:password).should_not be_nil
+      u.errors[:name].should_not be_nil
     end.should_not change(User, :count)
   end
 
   it 'requires password confirmation' do
     lambda do
       u = create_user(:password_confirmation => nil)
-      u.errors.on(:password_confirmation).should_not be_nil
+      u.errors[:name].should_not be_nil
     end.should_not change(User, :count)
   end
 
   it 'requires email' do
     lambda do
       u = create_user(:email => nil)
-      u.errors.on(:email).should_not be_nil
+      u.errors[:name].should_not be_nil
     end.should_not change(User, :count)
   end
 
@@ -75,10 +75,10 @@ describe User do
      'hello.-_there@funnychar.com', 'uucp%addr@gmail.com', 'hello+routing-str@gmail.com',
      'domain@can.haz.many.sub.doma.in', 'student.name@university.edu'
     ].each do |email_str|
-      it "'#{email_str}'" do
+      it "should work for '#{email_str}'" do
         lambda do
           u = create_user(:email => email_str)
-          u.errors.on(:email).should be_nil
+          u.errors[:email].should be_empty
         end.should change(User, :count).by(1)
       end
     end
@@ -93,7 +93,7 @@ describe User do
       it "'#{email_str}'" do
         lambda do
           u = create_user(:email => email_str)
-          u.errors.on(:email).should_not be_nil
+          u.errors[:name].should_not be_nil
         end.should_not change(User, :count)
       end
     end
@@ -103,10 +103,10 @@ describe User do
     ['Andre The Giant (7\'4", 520 lb.) -- has a posse',
      '', '1234567890_234567890_234567890_234567890_234567890_234567890_234567890_234567890_234567890_234567890',
     ].each do |name_str|
-      it "'#{name_str}'" do
+      it "should work for '#{name_str}'" do
         lambda do
           u = create_user(:name => name_str)
-          u.errors.on(:name).should be_nil
+          u.errors[:name].should be_empty
         end.should change(User, :count).by(1)
       end
     end
@@ -118,7 +118,7 @@ describe User do
       it "'#{name_str}'" do
         lambda do
           u = create_user(:name => name_str)
-          u.errors.on(:name).should_not be_nil
+          u.errors[:name].should_not be_nil
         end.should_not change(User, :count)
       end
     end
@@ -133,28 +133,56 @@ describe User do
     User.authenticate('quentin', 'invalid_password').should be_nil
   end
 
-  if REST_AUTH_SITE_KEY.blank?
-    it "authenticates a user against a hard-coded old-style password" do
-      User.authenticate('old_password_holder', 'test').should == users(:old_password_holder)
-    end
-  else
-    it "doesn't authenticate a user against a hard-coded old-style password" do
-      User.authenticate('old_password_holder', 'test').should be_nil
+  describe "find or create from google openid fetch response" do
+
+    it "should generate name/email from the fetch response" do
+      fetch_response = mock()
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/contact/email').and_return("wilma@example.com")
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/namePerson/first').and_return("Wilma")
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/namePerson/last').and_return("Flintstone")
+
+      lambda {
+        user = User.find_or_create_from_google_openid(fetch_response)
+        user.login.should == "wilma"
+        user.name.should == "Wilma Flintstone"
+        user.email.should == "wilma@example.com"
+        user.should be_valid
+      }.should change(User, :count).by(1)
     end
 
-    desired_encryption_expensiveness_ms = 0.1
-    it "takes longer than #{desired_encryption_expensiveness_ms}ms to encrypt a password" do
-      test_reps = 100
-      start_time = Time.now; test_reps.times{ User.authenticate('quentin', 'monkey'+rand.to_s) }; end_time = Time.now
-      auth_time_ms = 1000 * (end_time - start_time)/test_reps
-      auth_time_ms.should > desired_encryption_expensiveness_ms
+    it "handle blank names" do
+      fetch_response = mock()
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/contact/email').and_return("wilma@example.com")
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/namePerson/first').and_return("")
+      fetch_response.should_receive(:get_single).once.with('http://axschema.org/namePerson/last').and_return("")
+
+      lambda {
+        user = User.find_or_create_from_google_openid(fetch_response)
+        user.login.should == "wilma"
+        user.name.should == ""
+        user.email.should == "wilma@example.com"
+        user.should be_valid
+      }.should change(User, :count).by(1)
     end
+
+    it "should retrieve a user if they already exist" do
+      User.create!(:login => "wilma", :email => "wilma@example.com", :name => "Wilma Flintstone", :password => "password", :password_confirmation => "password")
+      fetch_response = mock()
+      fetch_response.should_receive(:get_single).with('http://axschema.org/contact/email').and_return("wilma@example.com")
+      fetch_response.should_receive(:get_single).with('http://axschema.org/namePerson/first').and_return("Wilma")
+      fetch_response.should_receive(:get_single).with('http://axschema.org/namePerson/last').and_return("Flintstone")
+
+      lambda {
+        User.find_or_create_from_google_openid(fetch_response)
+      }.should_not change(User, :count)
+    end
+
   end
 
   protected
 
   def create_user(options = {})
-    record = User.new({ :login => 'quire', :email => 'quire@example.com', :password => 'quire69', :password_confirmation => 'quire69' }.merge(options))
+    record = User.new({:login => 'quire', :email => 'quire@example.com', :password => 'quire69', :password_confirmation => 'quire69'}.merge(options))
     record.save
     record
   end
